+++ project/libs/shared/core/src/index.ts
@@ -1,69 +1,10 @@
 export { Entity } from './lib/base/entity';
 
-export { Token } from './lib/interfaces/token.interface';
-export { TokenPayload } from './lib/interfaces/token-payload.interface';
-export { PasswordToken } from './lib/interfaces/password-token.interface';
-export { JwtToken } from './lib/interfaces/jwt-token.interface';
-export { RefreshTokenPayload } from './lib/interfaces/refresh-token-payload.interface.ts';
-
-export { UserRole } from './lib/types/user-role.enum';
-export { User } from './lib/types/user.interface';
-export { AuthUser } from './lib/types/auth-user.interface';
-
-export { StorableEntity } from './lib/interfaces/storable-entity.interface';
-export { EntityFactory } from './lib/interfaces/entity-factory.interface';
-
-export { PaginationResult } from './lib/interfaces/pagination.interface';
-export { SortDirection } from './lib/interfaces/sort-direction.interface';
-export { SortBy } from './lib/interfaces/sort-by.interface';
-
-export { Comment } from './lib/types/blog/comment.interface';
-export { Like } from './lib/types/blog/like.interface';
-export { Tag } from './lib/types/blog/tag.interface';
-
-export { Post } from './lib/types/blog/posts/post.interface';
-export { PostType } from './lib/types/blog/posts/post-type.enum';
-export { PostStatus } from './lib/types/blog/posts/post-status.enum';
-export * from './lib/dto/blog/content';
-
-export { BasePostContent } from './lib/types/blog/posts/content/base-post-content.interface';
-export { LinkPostContent } from './lib/types/blog/posts/content/link-post-content.interface';
-export { PhotoPostContent } from './lib/types/blog/posts/content/photo-post-content.interface';
-export { QuotePostContent } from './lib/types/blog/posts/content/quote-post-content.interface';
-export { TextPostContent } from './lib/types/blog/posts/content/text-post-content.interface';
-export { VideoPostContent } from './lib/types/blog/posts/content/video-post-content.interface';
-
-export { CreatePostDto } from './lib/dto/blog/create-post.dto';
-export {
-  CreateUserDto,
-  CreateUserDtoWithAvatarFile,
-} from './lib/dto/users/create-user.dto';
-export { LoginUserDto } from './lib/dto/users/login-user.dto';
-export { RecoveryEmailDto } from './lib/dto/users/recovery-email.dto';
-
-export { PostRdo } from './lib/rdo/blog/post.rdo';
-export { UploadedFileRdo } from './lib/rdo/file-vault/uploaded-file.rdo';
-export { LoggedUserRdo } from './lib/rdo/users/logged-user.rdo';
-export { UserRdo } from './lib/rdo/users/user.rdo';
-
-export { File } from './lib/types/file-vault/file.interface';
-export { StoredFile } from './lib/types/file-vault/stored-file.interface';
-
-export { Subscriber } from './lib/types/notify/subscriber.interface';
-export { RabbitRouting } from './lib/types/notify/rabbit-routing.enum';
-export { RabbitExchange } from './lib/types/notify/rabbit-exchange.enum';
-export { RabbitQueue } from './lib/types/notify/rabbit-queue.enum';
-
-export { ChangeSubscriberPasswordDto } from './lib/dto/notify/change-password.dto';
-export { CreateSubscriberDto } from './lib/dto/notify/create-subscriber.dto';
-export { DeleteSubscriberDto } from './lib/dto/notify/delete-subscriber.dto';
-
-export { NotifyRecoveryEmailDto } from './lib/dto/notify/recovery-email.dto';
-export { RecoveryEmailRdo } from './lib/rdo/users/recovery-email.rdo';
-
-export { ChangePasswordDto } from './lib/dto/users/change-password.dto';
-export { ChangePasswordRdo } from './lib/rdo/users/change-password.rdo';
-
-export { RefreshUserRdo } from './lib/rdo/users/refresh-user.rdo';
-
-export { CronTime } from './lib/types/cronTime.enum';
+// DTO
+export * from './lib/dto/index';
+// RDO
+export * from './lib/rdo/index';
+// Interfaces
+export * from './lib/interfaces/index';
+// Enums
+export * from './lib/enums/index';
