+++ project/libs/shared/core/src/lib/dto/index.ts
@@ -1,0 +1,17 @@
+// Posts
+export { CreatePostDto } from './blog/create-post.dto';
+export * from './blog/content/index';
+// Users
+export {
+  CreateUserDto,
+  CreateUserDtoWithAvatarFile,
+} from './users/create-user.dto';
+export { LoginUserDto } from './users/login-user.dto';
+export { RecoveryEmailDto } from './users/recovery-email.dto';
+export { ChangeSubscriberPasswordDto } from './notify/change-password.dto';
+export { ChangePasswordDto } from './users/change-password.dto';
+
+// Notify
+export { NotifyRecoveryEmailDto } from './notify/recovery-email.dto';
+export { CreateSubscriberDto } from './notify/create-subscriber.dto';
+export { DeleteSubscriberDto } from './notify/delete-subscriber.dto';
