+++ project/libs/shared/core/src/lib/enums/index.ts
@@ -1,0 +1,10 @@
+export { CronTime } from './cron/cronTime.enum';
+// Users
+export { UserRole } from './users/user-role.enum';
+// Posts
+export { PostType } from './posts/post-type.enum';
+export { PostStatus } from './posts/post-status.enum';
+// Rabbit
+export { RabbitRouting } from './rabbit/rabbit-routing.enum';
+export { RabbitExchange } from './rabbit/rabbit-exchange.enum';
+export { RabbitQueue } from './rabbit/rabbit-queue.enum';
