+++ project/libs/shared/core/src/lib/enums/rabbit/rabbit-exchange.enum.ts
