+++ project/libs/shared/core/src/lib/enums/rabbit/rabbit-queue.enum.ts
