+++ project/libs/shared/core/src/lib/enums/rabbit/rabbit-routing.enum.ts
