+++ project/libs/shared/core/src/lib/interfaces/base/entity-factory.interface.ts
