+++ project/libs/shared/core/src/lib/interfaces/base/pagination.interface.ts
