+++ project/libs/shared/core/src/lib/interfaces/base/storable-entity.interface.ts
