+++ project/libs/shared/core/src/lib/interfaces/blog/posts/content/base-post-content.interface.ts
