+++ project/libs/shared/core/src/lib/interfaces/blog/posts/content/link-post-content.interface.ts
