+++ project/libs/shared/core/src/lib/interfaces/blog/posts/content/photo-post-content.interface.ts
