+++ project/libs/shared/core/src/lib/interfaces/blog/posts/content/quote-post-content.interface.ts
