+++ project/libs/shared/core/src/lib/interfaces/blog/posts/content/text-post-content.interface.ts
