+++ project/libs/shared/core/src/lib/interfaces/blog/posts/content/video-post-content.interface.ts
