+++ project/libs/shared/core/src/lib/interfaces/file-vault/file.interface.ts
