+++ project/libs/shared/core/src/lib/interfaces/file-vault/stored-file.interface.ts
