+++ project/libs/shared/core/src/lib/interfaces/index.ts
@@ -1,0 +1,39 @@
+// Base
+export { StorableEntity } from './base/storable-entity.interface';
+export { EntityFactory } from './base/entity-factory.interface';
+
+export { PaginationResult } from './base/pagination.interface';
+export { SortDirection } from './base/sort-direction.interface';
+export { SortBy } from './base/sort-by.interface';
+
+// Posts
+export { Post } from './blog/posts/post.interface';
+
+export { BasePostContent } from './blog/posts/content/base-post-content.interface';
+export { LinkPostContent } from './blog/posts/content/link-post-content.interface';
+export { PhotoPostContent } from './blog/posts/content/photo-post-content.interface';
+export { QuotePostContent } from './blog/posts/content/quote-post-content.interface';
+export { TextPostContent } from './blog/posts/content/text-post-content.interface';
+export { VideoPostContent } from './blog/posts/content/video-post-content.interface';
+
+export { Comment } from './blog/comment.interface';
+export { Like } from './blog/like.interface';
+export { Tag } from './blog/tag.interface';
+
+// FileVault
+export { File } from './file-vault/file.interface';
+export { StoredFile } from './file-vault/stored-file.interface';
+
+// Notify
+export { Subscriber } from './notify/subscriber.interface';
+
+// Users
+export { User } from './users/user.interface';
+export { AuthUser } from './users/auth-user.interface';
+
+// Tokens
+export { Token } from './users/tokens/token.interface';
+export { TokenPayload } from './users/tokens/token-payload.interface';
+export { PasswordToken } from './users/tokens/password-token.interface';
+export { JwtToken } from './users/tokens/jwt-token.interface';
+export { RefreshTokenPayload } from './users/tokens/refresh-token-payload.interface.ts';
