+++ project/libs/shared/core/src/lib/interfaces/notify/subscriber.interface.ts
