+++ project/libs/shared/core/src/lib/interfaces/users/tokens/jwt-token.interface.ts
