+++ project/libs/shared/core/src/lib/interfaces/users/user.interface.ts
@@ -1,4 +1,4 @@
-import { UserRole } from './user-role.enum';
+import { UserRole } from '../../enums/users/user-role.enum';
 
 export interface User {
   id?: string;
