+++ project/libs/shared/core/src/lib/rdo/index.ts
@@ -1,0 +1,12 @@
+// Users
+export { UserRdo } from './users/user.rdo';
+export { LoggedUserRdo } from './users/logged-user.rdo';
+export { RecoveryEmailRdo } from './users/recovery-email.rdo';
+export { ChangePasswordRdo } from './users/change-password.rdo';
+export { RefreshUserRdo } from './users/refresh-user.rdo';
+
+// Posts
+export { PostRdo } from './blog/post.rdo';
+
+// FileVault
+export { UploadedFileRdo } from './file-vault/uploaded-file.rdo';
